requirements.txt
torch
torchvision
numpy
scipy
einops
PyYAML
pytest
pandas
opencv-python
